requirements.txt
# GREENWIRE Static Distribution Requirements
# All dependencies pinned to specific versions for reproducible builds

# Core Python dependencies
cryptography==46.0.1
pycryptodome==3.23.0
pycryptodomex==3.23.0

# NFC and Smartcard libraries
pyscard==2.0.0
nfcpy==1.0.4
ndeflib==0.3.3
pyusb==1.3.1
libusb1==3.3.1
hidapi==0.14.0.post4
hid==1.0.8

# APDU Fuzzing dependencies
intelhex==2.3.0

# Advanced EMV and APDU libraries
emv==1.0.14
pyemv==1.5.0

# Hardware interface libraries (Windows-specific)
pyudev==0.24.3; sys_platform != "win32"
pywin32==311; sys_platform == "win32"
pywin32-ctypes==0.2.3; sys_platform == "win32"

# Networking and communication
requests==2.32.5
urllib3==2.5.0
websockets==15.0.1
paramiko==4.0.0
pyserial==3.5
beautifulsoup4==4.12.3

# Data processing and encoding
pyyaml==6.0.2
qrcode==8.2
pillow==11.3.0
numpy==2.3.2

# Security and encryption
bcrypt==4.3.0
pynacl==1.5.0
rsa==4.9.1
passlib==1.7.4
python-pkcs11==0.9.0

# Binary analysis (kept minimal; reverse-engineering tools excluded)
capstone==5.0.6
keystone-engine==0.9.2
pefile==2023.2.7

# Standard library extensions
python-dateutil==2.9.0.post0
six==1.17.0
colorama==0.4.6
click==7.1.2

# Menu system
windows-curses==2.4.1; sys_platform == "win32"

# Development and testing (optional)
pytest==8.4.1
pytest-cov==6.2.1
coverage==7.10.6
flake8==7.3.0
autopep8==2.3.2